Temp calculation from image row 321
This diff appears to extract the temp data for automated logging.  I've used it to automate thermal image capture and logging of user data based on a prox card swipe for upload to aws services. Have tested it and there is sometimes a 0.1 degree variance between temp that is displayed in the image and the temp that is output in the print statement due to the calculation being done on a different frame than is captured for image viewing. Thanks for the code, and I hope this helps others. I can share my HID access key fob automation code as well if it would help anyone. Here is a video showing the integration. https://youtu.be/h5mQt1Ct9b0